Conflicts:
	client/static/app/modules/play/templates/play-new.tpl.html

+++ server/modules/plays/plays.controller.js
@@ -14,7 +14,9 @@ module.exports.joinPlay = joinPlay;
 module.exports.leavePlay = leavePlay;
 
 function getPlays(req, res) {
-  PlayModel.findByDate(req.query.startDate, req.query.endDate)
+  PlayModel.findByDateAndCity(
+    req.query.startDate, req.query.endDate, req.query.city, req.query.page
+  )
   .then(function(data) {
     res.status(200).json(data);
   }, function(err) {
@@ -44,6 +46,8 @@ function getPlay(req, res) {
 
 function createPlay(req, res) {
   var dataFields = getRequestDataFields(req);
+  dataFields.creator = req.user._id;
+
   PlayModel.create(dataFields)
   .then(function(data) {
     res.status(200).json(data);
@@ -53,8 +57,15 @@ function createPlay(req, res) {
 }
 
 function deletePlay(req, res) {
-  PlayModel.findByIdAndRemove(req.params.playId).exec()
-  .then(function() {
+  PlayModel.findById(req.params.playId).exec()
+  .then(function(play) {
+    if (!play.creator.equals(req.user._id)) {
+      throw "Only owner can delete play";
+    } else {
+      return PlayModel.findByIdAndRemove(req.params.playId).exec();
+    }
+  })
+  .then(function(){
     res.status(200).json({});
   }, function(err) {
     res.status(500).json({error: err});
@@ -143,7 +154,8 @@ function getRequestDataFields(req) {
     name: req.body.name,
     playersMin: req.body.playersMin,
     playersMax: req.body.playersMax,
-    where: req.body.where,
+    city: req.body.city,
+    address: req.body.address,
     when: req.body.when,
     game: req.body.gameId,
     description: req.body.description
